ServerResources changed

+++ elocatorWeb/src/feri/rvir/elocator/rest/resource/tracking/TrackingServerResource.java
@@ -1,6 +1,6 @@
 package feri.rvir.elocator.rest.resource.tracking;
 
-import java.util.ArrayList;
+import java.util.List;
 
 import org.restlet.resource.ServerResource;
 
@@ -15,15 +15,16 @@ public class TrackingServerResource extends ServerResource implements TrackingRe
 	
 	
 	@Override
-	public Tracking retrieve() {
+	public List<Tracking> retrieve() {
 		System.out.println("RETRIEVE TrackingServerResource");
 		String username=(String)getRequest().getAttributes().get("username");
 		System.out.println(username);
-		// TODO pridobi objekt Tracking glede na username uporabnika
-		ArrayList<User> userList=new ArrayList<User>();
-		userList.add(new User("usernameExample","passwordExample"));
-		userList.add(new User("usernameExample","passwordExample"));
-		return null;
+
+		User u = udao.getUser(username);
+		
+		if (u == null) return null;
+		List<Tracking> trackings = tdao.getTrackingsByUser(u.getKey());
+		return trackings;
 	}
 
 	@Override
@@ -33,25 +34,14 @@ public class TrackingServerResource extends ServerResource implements TrackingRe
 	}
 
 	@Override
-	public void store(String usernameTracker, String usernameBeingTracked) {
-		System.out.println("STORE2 TrackingServerResource");
-	}
-
-	@Override
-	public void remove(String username) {
-		System.out.println("REMOVE TrackingServerResource");
-		User u = udao.getUser(username);
-		// remove tracking by tracker
-		tdao.deleteTrackByUserKey(u.getKey());
-	}
-
-	@Override
 	public void remove(String usernameTracker, String usernameBeingTracked) {
 		System.out.println("REMOVE2 TrackingServerResource");
 		User tracker = udao.getUser(usernameTracker);
 		User child = udao.getUser(usernameBeingTracked);
+		
+		if (tracker == null || child == null) return;
+		
 		tdao.deleteUserFromBeingTracked(tracker.getKey(),child.getKey());
-
 	}
 
 }
